Move developer docs to CONTRIBUTING.md
Expand upon README for first-time users, installing the project,
and making their first GraphQL query.

The steps to download and build a private Go project are silly.

Once this project is public it should be as easy as:

`$ go install github.com/circleci/circleci-cli`

Without having to stress about distribution.