Extract config into it's own package
This has an awkward of side-effect of referencing the internal config
struct as config.Config -- welcome to naming ideas!

An alternative would be to store a config of type config.Config in
cmd, and instead of calling config.Init from OnInitialize,
we could have a init function in cmd that sets up the struct.