Move existing integration test to cmd package
So that we can organise the tests per sub-command/verb. The current
`--help` test belongs to functionality in `root.go`.

I've left the remaining contents of `circleci_cli_suite_test.go` so that
you can run `ginkgo -r` from the root of the repo. The `main` package
should be so simple that we don't need any specs there though.