Allow overriding config with flag and honor that during setup
Before this change the behavior was _always_ generate config at the
default path. This was to avoid confusion if you ran the command with
the flag one time and then not the next you would end up with more
than one config.

I think this behavior is least surprising that we would generate the
config at the given path if it didn't already exist.