Extend filetree tests
To demonstrate nested directories and files. I looked at using
`github.com/Flaque/filet` to create temp dirs/files but it needs a
concrete implementation of `*testing.T` which isn't compatible with
Ginkgo:

- http://onsi.github.io/ginkgo/#using-other-matcher-libraries
- https://godoc.org/github.com/onsi/ginkgo#GinkgoT

It might be worth raising a pull request upstream and switching later,
but the now we can use `ioutil.WriteFile` since we only append to the
file in one place.

I've also added `HaveLen` matchers to check that there aren't more items
in the slice than we are making explicit assertions against.