Add support for "orb.yml" case to merge contents under the parent
This case occurs only under the following situations:

* we're on a leaf node that is a file AND
* it's siblings are all _only_ directories

We'll probably need to tweak this a little in the case we want to
support merging other files at this level of the tree.

`usage.yml` maybe one such case.

This change also identified a case where this type of file,
let's call it the "root file" can appear at the top-level of a tree.

In this case the contents of the root file are nested under a key by
the same name of the path which we started.

At the moment, this means in our tests we'd add a key named after the
temporary file path we constructed for our test environment. :/