Have my :cake: and add a test for the default path on orb validate
By allowing us to specify the path when creating a temp file, we can
write directly to the working directory of the test executable.

This resolves my comments on 74deab2.

Also make sure temp files and folders are properly disposed of,
this is something we're leaving up to who opened them first.

Since using `defer` properly won't work because context is left
between tests -- we should use `AfterEach` to close them ourselves.