Ensure exec will work
We weren't using command properly and realized that although golang
specifies the second argument to syscall.Exec as argv0, the first
element in that slice should be the name of the executable.

Actually, the first argument includes the binary, but we still have to
include something in that place.. it could be anything!

This change also ensures flag parsing is disabled for proxied
commands, updates some of the messaging, and adds some docs.