Enable snapcraft deploys
There is a slight change in the `circleci version` command to print a
valid version signifier compatible with snapcraft's syntax.

Instead of printing `0.1.X (commit)` we print `0.1.X+commit`.

This was chosen over `0.1.X-commit` as it is the example from their
documentation: https://docs.snapcraft.io/build-snaps/syntax#version

Here was the error that occurred before this change:

> Issues while validating snapcraft.yaml: The 'version' property does not match the required schema: '0.1.1160 (ed5fdaa)' is not a valid snap version. Snap versions consist of upper- and lower-case alphanumeric characters, as well as periods, colons, plus signs, tildes, and hyphens. They cannot begin with a period, colon, plus sign, tilde, or hyphen. They cannot end with a period, colon, or hyphen.