Use @dnephin's suggested PositionalArgs and FormatPositionalArg fns
While both cmd and md_docs packages use these functions for printing
positional arguments, putting them in cmd would result in a circular
import dependecy. So it was necessary to put these functions inside
the md_docs package for now.

It might make sense to have general "help" package, or something,
that would be we could put these and the template, etc.

Saving that for a later refactor.