API should return serialized Orb object
Since we use the same struct for both deserializing the API response
as we do for serializing back into JSON for `orb list`, we have to
maintain the same mapping the API gives us.

This means when we serialize an orb from the API it will have a
"version" field as well as "versions", we've discussed possibly
renaming this field to "highest_version" or similar.

Lastly, if there is a way to return the name and namespace (string) of
an orb in a query, printing the response would get simpler.

So we don't have to reuse `args` in the success message.