Improve error and response handling
This change allows our API code to have full control over the response
body in our GraphQL client, as well as expose all internal response
structures that include errors in-band and out as idomatic function
return values, as golang intended.

Before this change, the client had an internal structure for
representing the response body from a GraphQL server. Exposing this
interface to consumers has the trade-off of having to deal with the
data wrapping in our API code, but means we know exactly what the
server is returning.

The benefit of this change is that we can easily identify where errors
are returned in the various responses and the API code can determine
how to handle them contextually. Since they are returned like normal
errors as function return values, the caller don't have to care about
where the errors come from. In this case cobra commands, whether it's
a connection error, something went wrong inside go, or a user-land
error that was returned from the GraphQL API.

Having established a pattern for response and error handling
throughout our API, going forward we should be able to identify which
error messages we want to refactor and implement a template for
displaying GQL errors.