Surface out-of-band GraphQL errors for `namespace create`
This catches an error when trying to find an org that doesn't exist:

```
$ circleci namespace create asdf github asdf
Error: Must have member permission.: the organization 'asdf' under 'GITHUB' VCS-type does not exist. Did you misspell the organization or VCS?
exit status 255
```

We've also made sure the in-band errors for createNamespace mutation
are preferred over out-of-band errors and runtime errors.