Drop the prompt from update and make default non-interactive
This commit addresses feedback from @ndintenfass, @dnephin, and @bear.

Minimal test cases have been added for dry-run and install which
required some mocking of the GitHub releases API while trying to
reduce the amount of testing the client libraries or hitting network.

The main reason for doing so was that I was getting rate-limited while
writing these tests!