Move help flag checking into runExecute
It's worthwhile to test for the various cases of a user calling

"circleci local execute --help"

However, testing for this when the behavior is in the RunE function
passed to the Cobra command requires a lot of ceremony to set up a mock
cobra command. It makes more sense to perform this check inside
"runExecute".

As for printing the help text, there is help text provided with "docker
run circleci/picard build --help". There's also help text provided
by "circleci help local execute", so it seems more consistent CLI
experience to match the 2nd.