Tighten up Picard SHA extract regexp
On Fedora systems, the previous regular expression for extracting the
latest sha256 of the picard image was matching extra characters, which
resulted in garbage being written to
~/.circleci/build_agent_settings.json.

This caused the subsquent docker
run to fail complaining that the sha256 was not a valid repository or
tag, which was definitely the case.

This updated regular expression does the right thing on Fedora systems
now.

Tested on both Fedora and Ubuntu (the latter of which worked fine both
before and after this change)

Fixes #190