Allow an OrbElement to be a string pointer
We've added a bit of syntax to orbs: a command, executor, or job can be a
reference to another command/executor/job. For example:

```
orbs:
  foo:
    commands:
      foo:
        steps:
          - checkout
commands:
  myfoo: foo/foo
```

For now, I'm not bothering with anything other than making the above syntax not
log a warning about the orb being corrupted, mostly because I'm not at all sure
that the CLI is the correct place to be following these references and getting
parameter/description data. So the above syntax will work, but the description
will be blank and the parameters will be empty (regardless of the actual
content).