Catch all unsupported type errors when merging trees
This limitation was previously implicit, and we need to carefully
consider how to add more types as use-cases grow. For now only maps.

Here's an example trying to merge a scalar string file into the tree:

```
Error: Failed trying to marshal the tree to YAML : expected a map, got a `string` which is not supported at this time for "/home/zzak/work/circleci-cli/tmp/foo.yml"
exit status 255
```