Token validation should be command responsibility not HTTP client
This change moves the token validation out of our GraphQL/HTTP client
and into the cmd package itself.

Commands can opt-in to validateToken pre-hook function.

Refactor `orb` command to use a single persistent pre-hook for setting
up the client and options.

Combined client.NewUnauthorizedRequest and client.NewAuthorizedRequest
to a single function with a request.SetToken helper.