[CIRCLE-15483]: Catch all 400s from GitHub release api in `update`
We want to make sure we cover all bases and catch every 4** error
GitHub might return to us and display a verbose message for how to
treat these kinds of errors.

Since the majority of the the time this error occurs when the user
hits the API without setting a GITHUB_TOKEN, and results in a rate
limt for their IP address; these directions should help solve this problem.