Delegate the entire 'tests' subcommand to circleci-agent
These commands are intended to be run from inside a circleCI build, so the
official source for these commands is the agent.

We made some changes recently that were not reflected in the CLI because it
reimplemented some of the commands.

This change removes the duplicate implementation and calls out to the agent.