Improve package manager reporting
We want to better record the package manager that is in use.

Add the package manager to the `circleci version` command. This will
make it easier to see how the binary was installed.

Add the package manager to the user agent header. This allows us to
measure the popularity of different installation methods, which in turn
allows us to better focus our energy on the most popular package
managers.

Also report when Snapcraft is the package manager

This change also disables the auto-update when Snapcraft is in use,
since Snap manages the updates itself.