Skip auto-update check while running in CI.
The CLI will always check for an update when run, unless explicitly told
not to. This can lead to rate limiting on shared IPs, especially in
CircleCI jobs. This helps alleviate this issue by checking for the $CI
enviornment variable before attempting to check for an update on GitHub.

It might be better to throw a message that the CLI is outdated but
continue with the command, however, this would have to account for being
piped into configuration files.