contexts commands take an api.ClientInterface
Adds a setting to use rest API. When unset or `false`, test the Rest API
and fall back to using the graphql API if the test fails. When set to
`true`, use the Rest API without testing first.

The `setup` command will set `use_rest_api` to `true` unless it is
already set.

We Warn the user if we did an unnecessary rest test

we will never run the test when:
  - the host is CCI cloud
  - `use_rest_api` is explicitly set to `true`

In other cases, we will run the test. If it succeeds, we will print a
notice that the user could skip it in the future if desired (by toggling
the setting or running `setup`). Otherwise, if it fails, we fall back to
using the GraphQL client.

Examples:

with `host: https://k9s.sphereci.com`, the rest API is tested, it works,
the notice is printed, and we use the rest API.

with `host: https://circleci.com`, the rest API is NOT tested.

with `host: https://pal.circleci.com`, the rest API is tested, it
doesn't work