Warn user if we did an unnecessary rest test
we will never run the test when:
  - the host is CCI cloud
  - `use_rest_api` is explicitly set to `true`

In other cases, we will run the test. If it succeeds, we will print a
notice that the user could skip it in the future if desired (by toggling
the setting or running `setup`). Otherwise, if it fails, we fall back to
using the GraphQL client.

Examples:

with `host: https://k9s.sphereci.com`, the rest API is tested, it works,
the notice is printed, and we use the rest API.

with `host: https://circleci.com`, the rest API is NOT tested.

with `host: https://pal.circleci.com`, the rest API is tested, it works
-- OOPS, this should not happen, it shouldn't work