Naming and organization changes
A few changes:

Move `client` directory/package to `api/graphql`, and rename the package
`graphql`. I think this makes it much more clear that `api` is the
higher-level interface to CCI, while the package formerly known as
`client` is a lower-level graphql client.

This still isn't anywhere CLOSE to ideal - specifically, the `api`
directory is still frankly a mess:

```
api/
├── api.go
├── api_suite_test.go
├── api_test.go
├── context.go
├── context_graphql.go
├── context_rest.go
├── context_test.go
└── graphql
    ├── client.go
    ├── client_test.go
    └── LICENSE
```

but overall I think it's better - the graphql bits are more contained,
and it's more clear that a new command using graphql should use the
`api` package, not the `graphql` package directly.

Along with this, I've renamed several things to make them more clear:

- `api.ClientInterface` -> `api.ContextInterface`
- `api.Client` -> `api.ContextRestClient`
- `api.NewClient` -> `api.NewContextRestClient`
- new `api.NewContextGraphqlClient` constructor, so callers don't need
  to reach down and create their own graphql.Client