Remove slack webhook configuration in CI
As a result of the CodeCov incident, we need to roll the SLACK_WEBHOOK environment variable, and t's
currently being used by a super outdated slack orb. Because this webhook was configured with an old
method, Bear and I couldn't detect exactly which of the 54 configurations the variable pointed to.

We are instead going to remove slack notification commands from all jobs and we'll add them back at
a later time.