Don't panic on bad semver versions
Homebrew versions may look like `0.1.1234_5` if there is a revision made
to the homebrew formula. This won't parse with semver.MustParse, so the
update check panics in this case.

This makes two changes. First, if we can't parse a version as semver,
return an error with normal go error handling, rather than using
`MustParse` and panicking. Second, change `rootCmdPreRun` to print
errors, then throw them away and continue along. This way, if the
check-for-upgrade piece of the CLI is broken, the CLI is still usable.