added error message to FollowProject
removed FollowProject param

removed param

added back param

added new error

Return an error on 4xx response codes.

So we don’t return misleading status codes.

This commit is a checkpoint to getting working non-zero response codes when following a project fails.

The end state we want is to replace the one off response code checking with the built-in support for the internal API client. This is to get us in a working state where a `255` is returned if following fails.

return error message when follow project fails

updated xcode version