Return correct error in ConfigQuery
During some testing, it was discovered that attempting to validate a
config while passing an invalid token fails with the following error:

```
$ circleci --token "invalid-token" config validate .circleci/config.yml -v
...
Error: config compilation request returned an error: %!w(<nil>)
```

The behavior would previously check if an error existed, but return a
different one. After fixing that, the error prints correctly.