PIPE-4877 Remove config validate test with invalid org ID
There has been a change to how configs are compiled that makes this test
invalid.

Previously we would allow any random owner ID to be passed in
when compiling certified orbs as these were enabled by default and
didn't require any checks with the provided owner ID.

To support an upcoming feature, now when compiling the config, all
provided owner IDs are checked and must exist to validate that the
organization allows that type of orb.

The CLI can continue to use no owner ID with public certified orbs as
tested in the above case.

https://github.com/CircleCI-Public/circleci-cli/blob/2766f0e710c8727e346e247f6c779079a79b010c/integration_tests/features/circleci_config.feature#L63-L85