Removes the now redundant homebrew deploy jobs from our CI pipeline
As part of my efforts to get config validation working with the newly
exposed pipeline values, I noted that the homebrew deploy job fails
due to this error:

```
Whoops, the ******** formula has its version update
pull requests automatically opened by BrewTestBot every ~3 hours!
We'd still love your contributions, though, so try another one
that's not in the autobump list:
```

As such, it makes sense to now remove this job and the shell script
and simplify our CI.

As part of this change, I've modified the README to include the new
instructions as to how a release is picked up by homebrew from our
main branch.