TEST-72 Update run command to autodetect envars
The testsuite plugin requires some envars to be able to run test
selection:

- `CIRCLE_URL` - to call the public API.
- `CIRCLE_TOKEN` - API auth.
- `CIRCLE_PROJECT_ID` - fetch impact data for a given project.
- `os.Environ` - allows running commands that depend on `os`.

These have been added to the `run` command, with the expectation that
other plugins will also require these envars.

The `projectRestClient` has been changed to exported so that its type
can be defined in a function arg.

The `settings.Config` now supports a stdout/stderr, providing a way
to test output of commands. The output is set to `os` when loading the
CLI config.